remove registerBoltWidgets from Widgets class
